+++ frites/conn/__init__.py
@@ -3,3 +3,4 @@ from .conn_covgc import conn_covgc  # noqa
 from .conn_dfc import conn_dfc  # noqa
 from .conn_transfer_entropy import conn_transfer_entropy  # noqa
 from .conn_fit import conn_fit  # noqa
+from .conn_utils import conn_reshape_undirected  # noqa