+++ frites/utils/__init__.py
@@ -1,4 +1,4 @@
 """Utility functions."""
-from .reshaping import conn_reshape_undirected  # noqa
+from .preproc import savgol_filter  # noqa
 from .window import define_windows, plot_windows  # noqa
 from .wrapper import jit  # noqa