+++ docs/source/whatsnew.rst
@@ -8,7 +8,7 @@ v0.3.5
 
 New Features
 ++++++++++++
-* New function for reshaping undirected arrays of connectivity (like DFC) :py:func:`frites.utils.conn_reshape_undirected` (:commit:`ffcae34`)
+* New function for reshaping undirected arrays of connectivity (like DFC) :py:func:`frites.conn.conn_reshape_undirected` (:commit:`ffcae34`)
 
 .. raw:: html